+++ CHANGELOG.md
@@ -50,6 +50,7 @@ For more information about this file see also [Keep a Changelog](http://keepacha
 - PEcAnRTM:
     - Exposed PROSPECT absorption coefficients and `gpm()` function ("generalized plate model"), facilitating experimentation with different absorption coefficients
     - Added `spectra` S3 class and methods for subsetting (e.g. `myspec[[400:700]]`), plotting (`plot()` and `matplot()`), and combining spectra by wavelength.
+    - Also added `tidy_spectra` function to convert `spectra` objects to long, tidy `tibble`s.
     - Added `resample` functions for quickly resampling spectra (and, more generally, vectors and functions) to different dimensions. 
     - `EDR` API has been revised. Setup has been refactored from EDR via new `setup_edr` function, which relies on the ED utilities (see `PEcAn.ED2` below), and the `EDR` function now focuses only on execution. Also, added new `params2edr` function to make it easy to convert complex EDR parameters list to flat parameter vector required by `invert_bt` (or other optimization functions).
     - New `invert_fieldspec` function that serves as a wrapper around `invert_bt` for more quickly performing PROSPECT inversions on field spectra.